requirements.txt
dash==2.10.2
dash_ag_grid==2.1.0
dash_bootstrap_components==1.4.1
matplotlib==3.7.2
numpy==1.24.3
pandas==2.0.2
plotly==5.14.1
python-dotenv==1.0.0
Requests==2.31.0
tensorflow==2.13.0
tensorflow_macos==2.13.0
tensorflow_probability==0.21.0
